refactor: move tests for stream and ops to separate modules
